+++ src/bl64-arc.bash
@@ -1,11 +1,39 @@
 #######################################
 # BashLib64 / Module / Functions / Manage archive files
 #
-# Version: 1.6.0
+# Version: 1.7.0
 #######################################
 
 #######################################
-# Tar wrapper with verbose, debug and common options
+# Unzip wrapper debug and common options
+#
+# Arguments:
+#   $@: arguments are passed as-is to the command
+# Outputs:
+#   STDOUT: command output
+#   STDERR: command stderr
+# Returns:
+#   command exit status
+#######################################
+function bl64_arc_run_unzip() {
+  bl64_dbg_lib_show_function "$@"
+  local verbosity='-qq'
+
+  bl64_check_command "$BL64_ARC_CMD_UNZIP" || return $?
+  bl64_msg_verbose_lib_enabled && verbosity='-q'
+  bl64_dbg_lib_command_enabled && verbosity=' '
+
+  # Ignore previous settings
+  unset UNZIP
+
+  # shellcheck disable=SC2086
+  "$BL64_ARC_CMD_UNZIP" \
+    $verbosity \
+    "$@"
+}
+
+#######################################
+# Tar wrapper debug and common options
 #
 # Arguments:
 #   $@: arguments are passed as-is to the command
@@ -17,14 +45,14 @@
 #######################################
 function bl64_arc_run_tar() {
   bl64_dbg_lib_show_function "$@"
-  local verbose=''
+  local debug=''
 
   bl64_check_command "$BL64_ARC_CMD_TAR" || return $?
-  bl64_dbg_lib_command_enabled && verbose="$BL64_ARC_SET_VERBOSE"
+  bl64_dbg_lib_command_enabled && debug="$BL64_ARC_SET_TAR_VERBOSE"
 
   # shellcheck disable=SC2086
   "$BL64_ARC_CMD_TAR" \
-    $verbose \
+    $debug \
     "$@"
 }
 
@@ -56,7 +84,10 @@ function bl64_arc_open_tar() {
     bl64_check_directory "$destination" ||
     return $?
 
-  cd "$destination" || return 1
+  bl64_msg_show_lib_task "$_BL64_ARC_TXT_OPEN_TAR ($source)"
+
+  # shellcheck disable=SC2164
+  cd "$destination"
 
   case "$BL64_OS_DISTRO" in
   ${BL64_OS_UB}-* | ${BL64_OS_DEB}-* | ${BL64_OS_FD}-* | ${BL64_OS_CNT}-* | ${BL64_OS_RHEL}-* | ${BL64_OS_ALM}-* | ${BL64_OS_OL}-*)
@@ -88,6 +119,46 @@ function bl64_arc_open_tar() {
     ;;
   *) bl64_check_show_unsupported ;;
   esac
+  status=$?
+
+  ((status == 0)) && bl64_fs_rm_file "$source"
+
+  return $status
+}
+
+#######################################
+# Open zip files and remove the source after extraction
+#
+# * Preserves permissions but not ownership
+# * Overwrites destination
+# * Ignore ACLs and extended attributes
+#
+# Arguments:
+#   $1: Full path to the source file
+#   $2: Full path to the destination
+# Outputs:
+#   STDOUT: None
+#   STDERR: tar or lib error messages
+# Returns:
+#   BL64_ARC_ERROR_INVALID_DESTINATION
+#   tar error status
+#######################################
+function bl64_arc_open_zip() {
+  bl64_dbg_lib_show_function "$@"
+  local source="$1"
+  local destination="$2"
+  local -i status=0
+
+  bl64_check_parameter 'source' &&
+    bl64_check_parameter 'destination' &&
+    bl64_check_directory "$destination" ||
+    return $?
+
+  bl64_msg_show_lib_task "$_BL64_ARC_TXT_OPEN_ZIP ($source)"
+  bl64_arc_run_unzip \
+    $BL64_ARC_SET_UNZIP_OVERWRITE \
+    -d "$destination" \
+    "$source"
   status=$?
 
   ((status == 0)) && bl64_fs_rm_file "$source"